+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### BREAKING CHANGES
 
+- `\Wizaplace\SDK\Translation\TranslationService::pushXliffCatalog` now uses a `$password` parameter instead of the apiClient authentication
+
 ### New features
 
 ### Bugfixes