+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### BREAKING CHANGES
 
+- Removed `\Wizaplace\SDK\Catalog\Product::getUrl`
+- Removed `\Wizaplace\SDK\Catalog\ProductSummary::getUrl`
+- Removed `\Wizaplace\SDK\Basket\BasketItem::getProductUrl`
+
 ### New features
 
 ### Bugfixes